heater-backend/requirements.txt
Flask==0.10.1
flask-restful==0.3.5
requests==2.7.0
PyYAML.Yandex==3.11.1
